+++ sz/src/sz_float.c
@@ -25,8 +25,8 @@ unsigned int optimize_intervals_float_1D(float *oriData, size_t dataLength, doub
 {	
 	size_t i = 0, radiusIndex;
 	float pred_value = 0, pred_err;
-	int *intervals = (int*)malloc(maxRangeRadius*sizeof(int));
-	memset(intervals, 0, maxRangeRadius*sizeof(int));
+	size_t *intervals = (size_t*)malloc(maxRangeRadius*sizeof(size_t));
+	memset(intervals, 0, maxRangeRadius*sizeof(size_t));
 	size_t totalSampleSize = dataLength/sampleDistance;
 	for(i=2;i<dataLength;i++)
 	{
@@ -69,8 +69,8 @@ unsigned int optimize_intervals_float_2D(float *oriData, size_t r1, size_t r2, d
 	size_t i,j, index;
 	size_t radiusIndex;
 	float pred_value = 0, pred_err;
-	int *intervals = (int*)malloc(maxRangeRadius*sizeof(int));
-	memset(intervals, 0, maxRangeRadius*sizeof(int));
+	size_t *intervals = (size_t*)malloc(maxRangeRadius*sizeof(size_t));
+	memset(intervals, 0, maxRangeRadius*sizeof(size_t));
 	size_t totalSampleSize = r1*r2/sampleDistance;
 	for(i=1;i<r1;i++)
 	{
@@ -116,8 +116,8 @@ unsigned int optimize_intervals_float_3D(float *oriData, size_t r1, size_t r2, s
 	size_t radiusIndex;
 	size_t r23=r2*r3;
 	float pred_value = 0, pred_err;
-	int *intervals = (int*)malloc(maxRangeRadius*sizeof(int));
-	memset(intervals, 0, maxRangeRadius*sizeof(int));
+	size_t *intervals = (size_t*)malloc(maxRangeRadius*sizeof(size_t));
+	memset(intervals, 0, maxRangeRadius*sizeof(size_t));
 	size_t totalSampleSize = (r1-1)*(r2-1)*(r3-1)/sampleDistance;
 	for(i=1;i<r1;i++)
 	{
